+++ src/entry_names_gbbasic.h
@@ -14,11 +14,11 @@
     DEF_PATTERN_ADDR(sig_gbbasic_actor_init_alpha3_at, 0x861A); // Bank 2: 0x461A
     DEF_PATTERN_ADDR(sig_gbbasic_actor_init_alpha4_at, 0x83C3); // Bank 2: 0x43C3
 
-    DEF_PATTERN_BUF(sig_gbbasic_actor_init_v11, AR_ARGS(0x19, 0xE5, 0x7D, 0xF8, 0x17, 0x77, 0xE1, 0x7C, 0xF8, 0x16, 0x32, 0x2A, 0x5F, 0x3A, 0x57, 0x1A, 0x22, 0x13, 0x1A, 0x77, 0xF8, 0x0E, 0x2A, 0x5F, 0x56));
-    DEF_PATTERN_ADDR(sig_gbbasic_actor_init_v11_at, 0x8650); // Bank 2: 0x4650
-    // Found in gbcompo version of reversi - may be post-1.1 version with engine changes
-    DEF_PATTERN_ADDR(sig_gbbasic_actor_init_v11_mod_at, 0x8242); // Bank 2: 0x4242
-
+    // TODO: Add support for masked patterns at address (accept start addr and len/-1)
+    DEF_PATTERN_BUF(sig_gbbasic_magic_v11, AR_ARGS(0x01, 0x01, 0x00, 0x00, 0x9B, 0x0B, 0xA5, 0x1C));
+    //                                             .MAJ, .MIN, .REV......, .MAGIC................ 
+    DEF_PATTERN_ADDR(sig_gbbasic_magic_v11_at, 0x20000); // Start of Fixed Bank 8: 0x20000
+    
 
     // Found by checking symbols in "kernel" which is a pre-compiled gb rom
     // bundled with the application. It's a region of stable (pre-compiled)
@@ -57,29 +57,24 @@
     // ld   d, [hl]
 
 
-    // V1.1 @ 02:4650 (midway through 02:440D _actor_init)
+    // V1.1 and maybe later
     //
-    // add  hl, de
-    // push hl
-    // ld   a, l
-    // ld   hl, sp+23
-    // ld   [hl], a
-    // pop  hl
-    // ld   a, h
-    // ld   hl, sp+22
-    // ldd  [hl], a
-    // ldi  a, [hl]
-    // ld   e, a
-    // ldd  a, [hl]
-    // ld   d, a
-    // ld   a, [de]
-    // ldi  [hl], a
-    // inc  de
-    // ld   a, [de]
-    // ld   [hl], a
-    // ld   hl, sp+14
-    // ldi  a, [hl]
-    // ld   e, a
-    // ld   d, [hl]    
+    // src/data/builtin.c
+    // 
+    // #pragma bank 8
+    //
+    // ...
+    //
+    // const unsigned char VERSION_MAJOR = GBBVM_VER_MAJOR; // Major ahead.
+    // const unsigned char VERSION_MINOR = GBBVM_VER_MINOR;
+    // const unsigned int VERSION_REVISION = GBBVM_VER_REVISION;
+    //
+    // // Head.
+    //
+    // const unsigned char MAGIC[] = {
+    //     0x9B, 0x0B, 0xA5, 0x1C
+    // };
+
+
 
 #endif // _ENTRY_NAMES_GBBASIC_H